// src/app/models/articulosProduccion.model.ts
export class ArticulosProduccionModel {
  idProduccion: number;
  idArticulo: number;
  idInventario: number;
  status: string;
  existencia: string;
  descripcion: string;
  fechaIngreso: string;
  idCategoria: number;
  idTalla: number;
  talla: string;
  ubicacion: string;
  categoria: string;
  imagen: string;
  precio: number;
  precio1: number;
  precio2: number;
  precio3: number;
  precio4?: number;
  precio5?: number;
  precio6?: number;
  precio7?: number;
  precio8?: number;
  precio9?: number;
  precio10?: number;
  sku: string;
  nombre: string;
  noCorte?: String;
  coste: number;
  stock?: number;
  idSubCategoria: number;
  subcategoria: string;
  CantMovimiento?: number;
  idUbicaciones?: number[];
  seguimiento: Boolean;
  color?: string;
  idUbicacion: any;
  max?: string;
  min?: string;
  stockApartado?: number;
  comentario?: String;
  error?: String;
  apartado?: String;
  constructor() {
    this.idUbicacion = 0;
    this.idArticulo = 0;
    this.idInventario = 0;
    this.status = '';
    this.descripcion = '';
    this.existencia = '0';
    this.fechaIngreso = '';
    this.idCategoria = 0;
    this.idSubCategoria = 0;
    this.idTalla = 0;
    this.categoria = '';
    this.talla = '';
    this.color = '';
    this.ubicacion = '';
    this.imagen = '';
    this.sku = '';
    this.max = '0';
    this.min = '0';
    this.CantMovimiento = 0;
    this.subcategoria = '';
    this.seguimiento = false;
    this.precio = 0;
    this.coste = 0;
    this.precio1 = 0;
    this.precio2 = 0;
    this.precio3 = 0;
    this.precio3 = 0;
    this.precio4 = 0;
    this.precio5 = 0;
    this.idUbicaciones = [];
    this.noCorte = '';
  }
}